Added fully functional user authentication via passport. Added role based authentication ie admin and front desk can view orders but other will be redirected to login. This has only been implemented in the order route with a possible plan to move it to its own route but for the time being take the passport code to your own routes and alter values for that route it redirect the role number to the correct route. Added registering a user which could be used in admin view but other than that it is mainly for testing purpose.
Added session storage for the user authentication via package.json.

So if using the previous version please update to this one as this fixes the issues with the system I was creating.

FIX
Add bootstrap for login.ejs.
Any errors with the authentication ie redirect not taking you to that page even with the correct user and password.
Add error page with the error code or add a nicer format.

Add logout button with working url.

Any questions email me.

RL