Implemented search functionality on orders pages that takes table number as a value this has used Op from sequelize to allow for like to be used in the query. Added a search page to display this information that is the same style as the orders page. Fixed issue with authentication login not redirecting to the correct area even with the session being stored.
Added the charged information from billing to the database along with the other tables that needed updating as a result of this data which is at a different route but still the same process.
charge.ejs now shows the correct data again.

Billing now sends all the data from the front end to the backend and to a Stripe for the payment processing.

RL